[Add, Edit, Delete] First version of package available on test pypi using this command `pip install -i https://test.pypi.org/simple/ susie`. Replaced setup.py with setup.cfg, added pyproject.toml, updated requirements.txt. Moved modules into one overall package directory in src. Added test notebook to track code examples.
