+++ requirements.txt
@@ -1,25 +1,36 @@
 appnope==0.1.3
 asttokens==2.4.0
 backcall==0.2.0
+certifi==2023.7.22
+charset-normalizer==3.3.0
 comm==0.1.4
 debugpy==1.8.0
 decorator==5.1.1
+docutils==0.20.1
 exceptiongroup==1.1.3
 executing==2.0.0
+idna==3.4
 importlib-metadata==6.8.0
 iniconfig==2.0.0
 ipykernel==6.25.2
 ipython==8.16.1
+jaraco.classes==3.3.0
 jedi==0.19.1
 jupyter_client==8.3.1
 jupyter_core==5.3.2
+keyring==24.2.0
+markdown-it-py==3.0.0
 matplotlib-inline==0.1.6
+mdurl==0.1.2
+more-itertools==10.1.0
 nest-asyncio==1.5.8
+nh3==0.2.14
 numpy==1.25.2
 packaging==23.1
 parso==0.8.3
 pexpect==4.8.0
 pickleshare==0.7.5
+pkginfo==1.9.6
 platformdirs==3.11.0
 pluggy==1.3.0
 prompt-toolkit==3.0.39
@@ -31,13 +42,19 @@ pytest==7.4.2
 pytest-mock==3.11.1
 python-dateutil==2.8.2
 pyzmq==25.1.1
+readme-renderer==42.0
+requests==2.31.0
+requests-toolbelt==1.0.0
+rfc3986==2.0.0
+rich==13.6.0
 scipy==1.11.2
 six==1.16.0
 stack-data==0.6.3
-Susie==0.0.1
 tomli==2.0.1
 tornado==6.3.3
 traitlets==5.11.2
+twine==4.0.2
 typing_extensions==4.8.0
+urllib3==2.0.6
 wcwidth==0.2.8
 zipp==3.17.0
