+++ setup.cfg
@@ -1,0 +1,27 @@
+[metadata]
+name = susie
+version = 1.0.2
+author = Malia Barker, Holly VanLooy
+author_email = maliabarker@icloud.com, hollyvanlooy@u.boisestate.edu
+description = Susie allows users to input exoplanet transit data and return information detailing possible tidal decay.
+long_description = file: README.md, LICENSE.txt
+long_description_content_type = text/markdown
+project_urls =
+    Bug Tracker = https://github.com/BoiseStatePlanetary/susie/issues
+    repository = https://github.com/BoiseStatePlanetary/susie/
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
+install_requires = 
+    numpy
+    scipy
+
+[options.packages.find]
+where = src
