--- setup.py
@@ -1,24 +1,0 @@
-import setuptools
-
-with open("README.md", "r") as fh:
-    long_description = fh.read()
-
-setuptools.setup(
-    name="susie",                           # This is the name of the package
-    version="0.0.1",                        # The initial release version
-    author="Malia Barker, Holly VanLooy",   # Full name of the author
-    author_email="maliabarker@icloud.com, hollyvanlooy@u.boisestate.edu",
-    description="Susie allows users to input exoplanet transit data and return information detailing possible tidal decay.",
-    long_description=long_description,      # Long description read from the the readme file
-    long_description_content_type="text/markdown",
-    packages=setuptools.find_packages(where="src"),                     # List of all python modules to be installed
-    classifiers=[
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-    ],                                      # Information to filter the project on PyPi website
-    python_requires='>=3.6',                # Minimum version requirement of the package
-    # py_modules=["ephemeris, transit_times"],                   # Name of the python package
-    package_dir={'':'src'},                 # Directory of the source code of the package
-    install_requires=['numpy', 'scipy']  # Install other dependencies if any
-)
